requirements.txt
# NatyaVeda Analyzer — Python Requirements
# Install with: pip install -r requirements.txt

# ── Core ML ──────────────────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# ── HuggingFace Ecosystem ────────────────────────────────────────────────────
transformers>=4.40.0        # VideoMAE-v2, RT-DETR, MotionBERT
huggingface-hub>=0.22.0
timm>=1.0.0                 # image backbone support
accelerate>=0.27.0          # multi-GPU / mixed precision

# ── TensorFlow (MoveNet Thunder via TF Hub) ───────────────────────────────────
tensorflow>=2.15.0
tensorflow-hub>=0.16.0

# ── MediaPipe (hand & body keypoints) ────────────────────────────────────────
mediapipe>=0.10.0

# ── MMPose Ecosystem (RTMW wholebody) ─────────────────────────────────────────
# Install separately via mim:
#   pip install openmim
#   mim install mmengine mmcv mmdet mmpose

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python>=4.9.0
Pillow>=10.0.0
imageio>=2.31.0
albumentations>=1.3.0       # image augmentation

# ── Video & Data Collection ───────────────────────────────────────────────────
yt-dlp>=2024.1.0
scenedetect[opencv]>=0.6.3
ffmpeg-python>=0.2.0

# ── Data & ML Utilities ───────────────────────────────────────────────────────
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0
einops>=0.7.0               # elegant tensor operations

# ── Training & Logging ────────────────────────────────────────────────────────
wandb>=0.16.0
tensorboard>=2.15.0

# ── Configuration & CLI ───────────────────────────────────────────────────────
pyyaml>=6.0
omegaconf>=2.3.0
rich>=13.0.0
tqdm>=4.65.0

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
